pyproject.toml
[tool.poetry]
name = "repl-nix-arthistorynavigator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = {extras = ["async"], version = "^3.0.3"}
requests = "^2.32.3"
aiohttp = "^3.10.5"
flask-login = "^0.6.3"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
psycopg2-binary = "^2.9.9"
openai = "^1.47.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
